Add command-line option to quickly adjust logging levels.
Adds --log-level command-line argument which expects a comma-separated
list of module:level pairs where module is the name as it appears in the
log output and level is an integer or level name as found in the
standard python logging module. The empty module name refers to the root
logger.