Disable source map generation
Since the source maps contain absolute file paths (substack/node-browserify#663)
the file version hashes change depending on where the project is checked
out (i.e. on different machines), even when the source files have not
changed. This causes unnecessary noise during commits.

We're not minifying at the moment anyway, so the source maps are
minimally helpful. They could be added back as part of "production"
build task in the future.