VEN detailed protocol conformity
This set of changes harmonizes the VEN with the conformance rules
in the OpenADR profile specification, including behavior that is
not exercised by the Kisensum VTN. The goal of these changes
is to ensure better compatibility with other VTNs, if and when this
VEN is called upon to interact with them.

Rule 3: Capture event priority and pass it through to the Control Agent.
Rule 22: Confirm whether the venID is present in an event's eiTarget.
Rule 42: eventResponse.requestID is set, so eiResponse.requestID is empty.
Rule 48: Report event protocol errors in oadrCreatedEvent eventResponse.
Rule 65: If an event with a startAfter gets canceled, randomize the cancel time.
Rule 301: Handle a metadata report request.
Rule 305: Report TELEMETRY_USAGE units as powerReal.
Rule 317: Pass requested report granularity to the Control Agent.
Rule 329: Include all reportRequestIDs in the oadrPendingReports list.
Rule 331: In TELEMETRY_USAGE, readingType should be Direct Read, not Mean.
Rule 333: Set both a min and a max oadrSamplingRate.
Rule 510: Report x-resourceStatus in a TELEMETRY_STATUS report.
Use generateDS v2.29.4: no more need to set original_tagname_.
Move some ControlAgent parameters to volttron config.
Refresh sample XML files.