Fixes unhandled exception when opening cursors for closed connections
If a connection was closed by the database driver due to an SSL, TCP, or
other error, the condition was never tested for and the connection was
never restored. When cursors were created for the connection, an
exception would be raised. The only remedy was to restart the agent.

This patch adds a test for closed connections before creating cursors
and attempts to automatically reopen the connection before proceeding.

See NN-308 and NN-105